Added legacy FindTypeTagName convention
We need to have this on for all RavenDB storages since all storages shared IDocumentStore.
The current legacy convention does not pluralize documents hence the reason we need this convention on otherwise we are not backwards compatible.

This means that there is no point of having a legacy mode flag since we need this setting regardless.
This legacy convention is only added if we create the IDocumentStore, if the user passes a IDocumentStore to us then we do not mess with existing conventions.