Temporary fix for one failing acceptancetest
Test: When_saga_is_mapped_to_complex_expr.Should_hydrate_and_invoke_the_existing_instance

Is failing with:

Exceptions = System.Reflection.TargetInvocationException: Exception has been thrown by the target of an invocation. ---> System.Exception: Correlating on saga properties not marked as unique is not safe due to the high risk for stale results. Please add a [Unique] attribute to the 'KeyValue' property on your 'TestSagaData' saga data class. If you still want to allow this please add .UsePersistence<RavenDBPersistence>().AllowStaleSagaReads() to your config